requirements.txt
PyYAML
asyncio
py-cord >= 2.3
SqlAlchemy >= 2.0.0
Flask
python-dotenv